+++ mypy/metastore.py
@@ -17,9 +17,7 @@ from abc import abstractmethod
 from collections.abc import Iterable
 from typing import TYPE_CHECKING, Any
 
-from mypy_extensions import i64
-
-from mypy.util import os_path_join
+from mypy.util import hash_path_stem, os_path_join
 
 if TYPE_CHECKING:
     # We avoid importing sqlite3 unless we are using it so we can mostly work
@@ -178,39 +176,6 @@ def connect_db(db_file: str, set_journal_mode: bool) -> sqlite3.Connection:
     return db
 
 
-def hash_path_stem(s: str) -> int:
-    """Hash the stem of a cache file path (everything before the first dot in the basename).
-
-    This is a combined stem-extraction + hash function optimized for mypyc compilation.
-    Uses only integer arithmetic, avoiding intermediate string allocations.
-    """
-    # First find end of stem (scanning backwards, stop at first dot after last separator)
-    i = len(s) - 1
-    end: i64 = i
-    while i >= 0:
-        c: i64 = ord(s[i])
-        if c == ord("/") or c == ord("\\"):
-            break
-        if c == ord("."):
-            end = i
-        i -= 1
-    # Calculate hash
-    hv: i64 = 123
-    i = end
-    while i >= 0:
-        c = i64(ord(s[i]))
-        hv = (hv * 33) ^ c
-        i -= 1
-    # Murmur3 finalizer for better bit avalanche (improves shard uniformity)
-    hv = (hv ^ (hv >> 32)) & 0xFFFFFFFF
-    hv ^= hv >> 16
-    hv = (hv * 0x85EBCA6B) & 0xFFFFFFFF
-    hv ^= hv >> 13
-    hv = (hv * 0xC2B2AE35) & 0xFFFFFFFF
-    hv ^= hv >> 16
-    return int(hv)
-
-
 class SqliteMetadataStore(MetadataStore):
     def __init__(
         self, cache_dir_prefix: str, set_journal_mode: bool = False, num_shards: int = 1
@@ -231,10 +196,7 @@ class SqliteMetadataStore(MetadataStore):
         else:
             for i in range(num_shards):
                 self.dbs.append(
-                    connect_db(
-                        os_path_join(cache_dir_prefix, f"cache.{i}.db"),
-                        set_journal_mode,
-                    )
+                    connect_db(os_path_join(cache_dir_prefix, f"cache.{i}.db"), set_journal_mode)
                 )
         # Track which shards have been written to since last commit.
         self.dirty_shards: set[int] = set()
